cosmetic

+++ rf_kernels/xorshift_plus.hpp
@@ -28,7 +28,7 @@ inline uint64_t rd64(std::random_device &rd)
 {
   uint32_t low32 = rd();
   uint32_t high32 = rd();
-  return (uint64_t(high32) << 32) | uint32_t(low32);
+  return (uint64_t(high32) << 32) | uint64_t(low32);
 }
